src/test/javascript/spec/app/entities/genre/genre-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';
import { ToastPlugin } from 'bootstrap-vue';

import * as config from '@/shared/config/config';
import GenreUpdateComponent from '@/entities/genre/genre-update.vue';
import GenreClass from '@/entities/genre/genre-update.component';
import GenreService from '@/entities/genre/genre.service';

import AlertService from '@/shared/alert/alert.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const i18n = config.initI18N(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.use(ToastPlugin);
localVue.component('font-awesome-icon', {});
localVue.component('b-input-group', {});
localVue.component('b-input-group-prepend', {});
localVue.component('b-form-datepicker', {});
localVue.component('b-form-input', {});

describe('Component Tests', () => {
  describe('Genre Management Update Component', () => {
    let wrapper: Wrapper<GenreClass>;
    let comp: GenreClass;
    let genreServiceStub: SinonStubbedInstance<GenreService>;

    beforeEach(() => {
      genreServiceStub = sinon.createStubInstance<GenreService>(GenreService);

      wrapper = shallowMount<GenreClass>(GenreUpdateComponent, {
        store,
        i18n,
        localVue,
        router,
        provide: {
          genreService: () => genreServiceStub,
          alertService: () => new AlertService(),
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.genre = entity;
        genreServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(genreServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.genre = entity;
        genreServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(genreServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });

    describe('Before route enter', () => {
      it('Should retrieve data', async () => {
        // GIVEN
        const foundGenre = { id: 123 };
        genreServiceStub.find.resolves(foundGenre);
        genreServiceStub.retrieve.resolves([foundGenre]);

        // WHEN
        comp.beforeRouteEnter({ params: { genreId: 123 } }, null, cb => cb(comp));
        await comp.$nextTick();

        // THEN
        expect(comp.genre).toBe(foundGenre);
      });
    });

    describe('Previous state', () => {
      it('Should go previous state', async () => {
        comp.previousState();
        await comp.$nextTick();

        expect(comp.$router.currentRoute.fullPath).toContain('/');
      });
    });
  });
});
